auth feature

+++ logic/buku/delete.php
@@ -2,27 +2,39 @@
 
 require_once __DIR__ . '/../../db-config.php';
 
-$id = $_POST['id'];
+if (isset($_POST['id'])) {
+  $id = $_POST['id'];
 
-$sql = "DELETE FROM books WHERE id = ?";
-$statement = $connection->prepare($sql);
-$statement->execute([$id]);
-$count = $statement->rowCount();
+  $sql = "DELETE FROM books WHERE id = ?";
+  $statement = $connection->prepare($sql);
+  $statement->execute([$id]);
+  $count = $statement->rowCount();
 
-if ($count == 1) {
-  $alert = <<<ALERT
+  if ($count == 1) {
+    $alert = <<<ALERT
     <script>
       alert('Delete Data Sucess!');
       window.location='../../book.php'
     </script>
   ALERT;
 
-  echo $alert;
-  exit();
+    echo $alert;
+    exit();
+  } else {
+    $alert = <<<ALERT
+    <script>
+      alert('Delete Data Gagal!');
+      window.location='../../book.php'
+    </script>
+  ALERT;
+
+    echo $alert;
+    exit();
+  }
 } else {
   $alert = <<<ALERT
     <script>
-      alert('Delete Data Gagal!');
+      alert('Data yang ingin di hapus tidak ditemukan!');
       window.location='../../book.php'
     </script>
   ALERT;
